[feat] make useSearchKeyInfos
